refactor: Relocate time-related utilities to a shared `utils` directory, update PostCSS and TypeScript configurations, and refine the Footer component.
