+++ components/TimestampModal.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
 import ModalButton from './modal/ModalButton';
-import { getCurrentTime, Timezones } from '../pages/api/functions/timeNow';
+import { getCurrentTime, Timezones } from '../utils/timeNow';
 import { store } from '../store/store';
 import ModalBase from './modal/ModalBase';
 import TableRow from './ui-elements/TableRow';
