+++ package.json
@@ -10,27 +10,28 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@headlessui/react": "^1.6.6",
-    "@heroicons/react": "^1.0.6",
-    "@reduxjs/toolkit": "^1.8.2",
-    "next": "^12.1.5",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "react-icons": "^4.4.0",
-    "react-redux": "^8.0.2",
-    "swr": "^1.3.0"
+    "@headlessui/react": "^2.2.9",
+    "@heroicons/react": "^2.2.0",
+    "@reduxjs/toolkit": "^2.11.2",
+    "next": "^16.2.1",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "react-icons": "^5.6.0",
+    "react-redux": "^9.2.0",
+    "swr": "^2.4.1"
   },
   "devDependencies": {
-    "@types/node": "17.0.25",
-    "@types/react": "18.0.6",
-    "@types/react-dom": "18.0.2",
-    "autoprefixer": "^10.4.5",
-    "eslint": "8.13.0",
-    "eslint-config-next": "12.1.5",
-    "postcss": "^8.4.12",
-    "tailwind-scrollbar": "^1.3.1",
-    "tailwindcss": "^3.0.24",
+    "@tailwindcss/postcss": "^4.2.2",
+    "@types/node": "^25.5.0",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "autoprefixer": "^10.4.27",
+    "eslint": "^10.1.0",
+    "eslint-config-next": "^16.2.1",
+    "postcss": "^8.5.8",
+    "tailwind-scrollbar": "^4.0.2",
+    "tailwindcss": "^4.2.2",
     "timezone": "^1.0.23",
-    "typescript": "4.6.3"
+    "typescript": "^6.0.2"
   }
 }
