+++ public/styles/globals.css
@@ -1,6 +1,12 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import "tailwindcss";
+
+@theme {
+  --font-nova-flat: "Nova Flat";
+  --font-proxima-nova: "Proxima Nova";
+}
+
+@plugin "tailwind-scrollbar";
+
 
 html,
 body {
